Move unlocking logic down to Home
