+++ src/features/home/Home.tsx
@@ -1,13 +1,27 @@
 import { Col, Row, Typography } from 'antd';
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 
 import { ListMarkets } from '../operator/Market/ListMarkets';
+import { UnlockModalForm } from '../walletUnlocker/UnlockModalForm';
+import { useIsReadyQuery } from '../walletUnlocker/walletUnlocker.api';
 
 import { DashboardPanelLeft } from './DashboardPanelLeft';
 import { DashboardPanelRight } from './DashboardPanelRight';
 
+const { Title } = Typography;
+
 export const Home = (): JSX.Element => {
-  const { Title } = Typography;
+  const { data: isReady } = useIsReadyQuery();
+  // UnlockWallet Modal
+  const [isUnlockWalletModalVisible, setIsUnlockWalletModalVisible] = useState(false);
+  const showUnlockWalletModal = () => setIsUnlockWalletModalVisible(true);
+  const handleUnlockWalletModalCancel = () => setIsUnlockWalletModalVisible(false);
+
+  useEffect(() => {
+    if (!isReady || (isReady.isInitialized && !isReady.isUnlocked)) {
+      showUnlockWalletModal();
+    }
+  }, [isReady]);
 
   return (
     <>
@@ -27,6 +41,11 @@ export const Home = (): JSX.Element => {
         Markets
       </Title>
       <ListMarkets />
+      <UnlockModalForm
+        closable={false}
+        handleUnlockWalletModalCancel={handleUnlockWalletModalCancel}
+        isUnlockWalletModalVisible={isUnlockWalletModalVisible}
+      />
     </>
   );
 };
